requirements.txt
aiohttp==3.8.4
PyYAML==6.0
requests==2.28.1
transformers==4.27.1
TensorFlow